% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclim.R
\name{bioclim}
\alias{bioclim}
\title{Bioclim model}
\usage{
bioclim(x, nq = 10)
}
\arguments{
\item{x}{A matrix of environmental values at presence locations}

\item{nq}{Number of quantiles to estimate bioclim envelopes.}
}
\value{
An object of class "bioclim"
}
\description{
This function fits a bioclim envelope model
}
