% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexhull.R
\name{convexhull}
\alias{convexhull}
\title{Convex hull model}
\usage{
convexhull(x)
}
\arguments{
\item{x}{A matrix of data of environmental values at observed locations}
}
\value{
An object of class "convexhull"
}
\description{
This function fits a convex hull model
}
