% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domain}
\alias{domain}
\title{Domain model}
\usage{
domain(x)
}
\arguments{
\item{x}{A matrix of data of environmental values at observed locations}
}
\value{
An object of class "domain"
}
\description{
This function fits a  domain model based on Gower's distance.
}
\details{
Distances D are converted to similarity index by the complement (1-D).
Based on Carpenter et al. (1993). Biodiversity Conservation 2, 667-680
}
