% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel}
\alias{kernel}
\title{Kernel model}
\usage{
kernel(x, ...)
}
\arguments{
\item{x}{A matrix of data of environmental values at observed locations}

\item{...}{Arguments to be passed to kernel (seel ?ks::kde)}
}
\value{
An object of class "kernel"
}
\description{
This function fits a kernel model
}
\details{
Kernel estimations can be computational intensive.
}
