% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R
\name{mahalanobis}
\alias{mahalanobis}
\title{Mahalanobis model}
\usage{
mahalanobis(x)
}
\arguments{
\item{x}{A matrix of data of environmental values at observed locations}
}
\value{
An object of class "mahalanobis"
}
\description{
This function fits a model based on mahalanobis distance.
Distance values are converted to p-values using chi-squared distribution
}
