% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal.R
\name{mvnormal}
\alias{mvnormal}
\title{Multivariate Normal Distribution model}
\usage{
mvnormal(x)
}
\arguments{
\item{x}{A matrix of data of environmental values at observed locations}
}
\value{
An object of class "mvnormal"
}
\description{
This function fits a Multivariate Normal Distribution model
}
