% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexhull.R
\name{plot.convexhull}
\alias{plot.convexhull}
\title{Plot Method for Convex Hull}
\usage{
\method{plot}{convexhull}(x, cols = 1:2, border = "red", pnt.col = "gray", add = FALSE, ...)
}
\arguments{
\item{x}{An object of class `convexhull`.}

\item{cols}{the columns indices of the data (variables) to use to plot. Only 2 values are used.}

\item{border}{The color of the polygon borders deppicting model extent,}

\item{pnt.col}{The color of the training data points (if add=FALSE).}

\item{add}{Boolean to add to current device or, if FALSE, to generate a new plot.}

\item{...}{Other plotting parameters to be passed (e.g. lwd, lty,...).}
}
\description{
This function plots the `convexhull` model.
}
