% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{plot.domain}
\alias{plot.domain}
\title{Plot Method for Domain}
\usage{
\method{plot}{domain}(
  model,
  cols = 1:2,
  contours = seq(0.9, 1, 0.01),
  border = "red",
  pnt.col = "gray",
  add = FALSE,
  ...
)
}
\arguments{
\item{model}{An object of class `domain`.}

\item{cols}{the columns indices of the data (variables) to use to plot. Only 2 values are used.}

\item{border}{The color of the polygon borders deppicting model extent,}

\item{pnt.col}{The color of the training data points (if add=FALSE).}

\item{add}{Boolean to add to current device or, if FALSE, to generate a new plot.}

\item{...}{Other plotting parameters to be passed (lwd, lty,...).}

\item{contouns}{Where to plot the extents on Gower's similarity values.}
}
\description{
This function plots the `domain` model.
}
