% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{plot.kernel}
\alias{plot.kernel}
\title{Plot Method for Kernel}
\usage{
\method{plot}{kernel}(
  x,
  cols = 1:2,
  contours = 10,
  border = "red",
  pnt.col = "gray",
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{A model object of class `kernel`.}

\item{cols}{the columns indices of the data (variables) to use to plot. Only 2 values are used.}

\item{contours}{How many contours to plot from kernel model.}

\item{border}{The color of the polygon borders deppicting model extent,}

\item{pnt.col}{The color of the training data points (if add=FALSE).}

\item{add}{Boolean to add to current device or, if FALSE, to generate a new plot.}

\item{...}{Other plotting parameters to be passed (lwd, lty,...).}
}
\description{
This function plots the `kernel` model.
}
