% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R
\name{plot.mahalanobis}
\alias{plot.mahalanobis}
\title{Plot Method for mahalanobis}
\usage{
\method{plot}{mahalanobis}(
  model,
  cols = 1:2,
  contours = c(1, 1.64, 1.96, 2.33),
  border = "red",
  pnt.col = "gray",
  add = FALSE,
  ...
)
}
\arguments{
\item{model}{An object of class `mahalanobis`.}

\item{cols}{the columns indices of the data (variables) to use to plot. Only 2 values are used.}

\item{contours}{A vector of contours to plot defined as standard deviations away from the mean value.}

\item{border}{The color of the polygon borders deppicting model extent,}

\item{pnt.col}{The color of the training data points (if add=FALSE).}

\item{add}{Boolean to add to current device or, if FALSE, to generate a new plot.}

\item{...}{Other plotting parameters to be passed (lwd, lty,...).}
}
\description{
This function plots the `mahalanobis` model.
}
