% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promodel.R
\name{plot.proniche}
\alias{plot.proniche}
\title{Plot method for proniche models}
\usage{
\method{plot}{proniche}(model, ...)
}
\arguments{
\item{model}{An object of class `proniche`.}

\item{...}{Additional graphical parameters passed to the underlying plot
method of the fitted model.}
}
\value{
A plot of the fitted niche model in environmental space.
}
\description{
Plot method for proniche models
}
