% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{predict.domain}
\alias{predict.domain}
\title{Predict Method for Domain}
\usage{
\method{predict}{domain}(object, newdata = NULL, type = "raw", ...)
}
\arguments{
\item{object}{A model object of class `domain`.}

\item{newdata}{New data for predictions in a form of matrix or coercible to matrix.  If NULL (default) predictions are given to training data.}

\item{type}{Type of response: either "raw" (default) for original values of model or "truncated" for niche truncation if defined.}

\item{...}{Additional arguments (currently ignored, included for consistency with the generic).}
}
\value{
A vector of predictions.
}
\description{
This function predicts values using a `domain` object.
}
