% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promodel.R
\name{predict.proniche}
\alias{predict.proniche}
\title{Predict method for proniche models}
\usage{
\method{predict}{proniche}(object, newdata = NULL, type = "raw", ...)
}
\arguments{
\item{object}{A model object of class `proniche`.}

\item{newdata}{A matrix, data frame, or SpatRaster containing the predictor
variables. If a SpatRaster is supplied, a SpatRaster is returned.}

\item{...}{Additional arguments to model prediction.}
}
\value{
A numeric vector or SpatRaster with predicted suitability values.
}
\description{
Predict method for proniche models
}
