% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexhull.R
\name{print.convexhull}
\alias{print.convexhull}
\title{Print Method for Convex Hull}
\usage{
\method{print}{convexhull}(x, ...)
}
\arguments{
\item{x}{An object of class `convexhull`.}

\item{...}{Additional arguments (currently ignored, included for consistency with the generic).}
}
\description{
Print simple information for object of class `convexhull`.
}
