% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{print.domain}
\alias{print.domain}
\title{Print Method for Domain}
\usage{
\method{print}{domain}(model)
}
\arguments{
\item{model}{An object of class `domain`.}
}
\description{
Print simple information for object of class `domain`.
}
