% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R
\name{print.mahalanobis}
\alias{print.mahalanobis}
\title{Print Method for mahalanobis}
\usage{
\method{print}{mahalanobis}(x, ...)
}
\arguments{
\item{x}{An object of class `mahalanobis`.}

\item{...}{Additional arguments (currently ignored, included for consistency with the generic).}
}
\description{
Print simple information for object of class `mahalanobis`.
}
