% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal.R
\name{print.mvnormal}
\alias{print.mvnormal}
\title{Print Method for Multivariate Normal Distribution model}
\usage{
\method{print}{mvnormal}(x)
}
\arguments{
\item{x}{An object of class `mvnormal`.}
}
\description{
Print simple information for object of class `mvnormal`.
}
