% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promodel.R
\name{truncate.proniche}
\alias{truncate.proniche}
\title{Truncate method for proniche models}
\usage{
\method{truncate}{proniche}(x, p_obs = 1, test_data = NULL)
}
\arguments{
\item{x}{An object of class `proniche`.}

\item{p_obs}{Proportion of occurrence observations to include in the truncated envelope, between 0 and 1.}

\item{test_data}{Optional data frame containing test values of the variables in 'con'}
}
\value{
The truncated fitted model in environmental space.
}
\description{
Truncate method for proniche models
}
