% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexhull.R
\name{convexhull}
\alias{convexhull}
\title{Convex hull model}
\usage{
convexhull(x, ...)
}
\arguments{
\item{x}{A matrix of environmental values at observed presence localities}

\item{...}{Optional additional arguments to pass to geometry::convhulln() -- e.g., try adding options="QJ" if you get an error due to insufficient number of unique points}
}
\value{
An object of class "convexhull"
}
\description{
This function fits a convex hull model
}
