% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqPlot.R
\name{freqPlot}
\alias{freqPlot}
\title{Frequency plot}
\usage{
freqPlot(
  vals,
  vars = NULL,
  type = "density",
  na.rm = TRUE,
  dup.rm = FALSE,
  verbosity = 2,
  ...
)
}
\arguments{
\item{vals}{data frame with the values of the variables (columns) at the presence localities (rows). Column names must match the names of the corresponding variables in 'vars'.}

\item{vars}{optional SpatRaster or data frame with the variables with which to visually compare the frequency of 'vals'. Must include the same variable names as `vals`.}

\item{type}{type of plot to produce: "density" or "boxplot". Case is ignored and partial argument matching is used (i.e., you can set e.g. type="Box")}

\item{na.rm}{logical (default TRUE) indicating whether to remove rows with missing values in 'vals'. Otherwise, some methods may produce an error or empty results.}

\item{dup.rm}{logical (default FALSE) indicating whether to remove rows with duplicate values for all variables in 'vals'.}

\item{verbosity}{numeric value indicating the amount of messages to display. The default is 2, for the maximum number of messages available.}

\item{...}{Optional additional arguments that can be passed to plot(), e.g. 'cex.main' or 'cex.axis'}
}
\value{
Invisibly returns `NULL`. The function is called for its side effect
  of producing a plot.
}
\description{
This function plots (in green) the frequency distribution of each variable at the presence localities, and optionally (in grey) also the frequency across a given study region, using either density or box plots.
}
\details{
This function provides a simple way to visualize the frequency distribution of one or more variables, optionally comparing the distribution of \code{vals} with \code{vars}. It supports two types of plots: density plots and boxplots. When \code{vars} is provided, the function overlays the comparison dataset onto the same plot.
}
\examples{
set.seed(123)
vals <- data.frame(var1 = rnorm(50), var2 = rnorm(50))
freqPlot(vals, type = "density")
}
