% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R
\name{mahalanobis}
\alias{mahalanobis}
\title{Mahalanobis model}
\usage{
mahalanobis(x)
}
\arguments{
\item{x}{A matrix of data of environmental values at observed locations}

\item{nq}{Number of quantiles to estimate bioclim envelopes.}
}
\value{
An object of class "mahalanobis"
}
\description{
This function fits a model based on mahalanobis distance.
Distance values are converted to similarity with (1 - M/max(M))
}
