% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclim.R
\name{predict.bioclim}
\alias{predict.bioclim}
\title{Predict Method for Bioclim}
\usage{
\method{predict}{bioclim}(model, newdata = NULL)
}
\arguments{
\item{model}{An object of class `bioclim`.}

\item{newdata}{New data for predictions in a form of matrix or conversible to matrix.  If NULL (default) predictions are given to training data.}
}
\value{
A vector of predictions.
}
\description{
This function predicts values using a `bioclim` object.
}
