% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexhull.R
\name{predict.convexhull}
\alias{predict.convexhull}
\title{Predict Method for Convex Hull}
\usage{
\method{predict}{convexhull}(object, newdata = NULL)
}
\arguments{
\item{object}{A model object of class `convexhull`.}

\item{newdata}{New data for predictions in a form of matrix or conversible to matrix.  If NULL (default) predictions are given to training data.}
}
\value{
A vector of predictions.
}
\description{
This function predicts values using a `convexhull` object.
}
