% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{predict.kernel}
\alias{predict.kernel}
\title{Predict Method for Kernel}
\usage{
\method{predict}{kernel}(model, newdata = NULL)
}
\arguments{
\item{model}{An object of class `kernel`.}

\item{newdata}{New data for predictions in a form of matrix or conversible to matrix.  If NULL (default) predictions are given to training data.}
}
\value{
A vector of predictions.
}
\description{
This function predicts values using a `kernel` object.
}
