% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R
\name{predict.mahalanobis}
\alias{predict.mahalanobis}
\title{Predict Method for Mahalanobis}
\usage{
\method{predict}{mahalanobis}(object, newdata = NULL)
}
\arguments{
\item{object}{An object of class `mahalanobis`.}

\item{newdata}{New data for predictions in a form of matrix or conversible to matrix.  If NULL (default) predictions are given to training data.}
}
\value{
A vector of predictions.
}
\description{
This function predicts values using a `mahalanobis` object.
}
