% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal.R
\name{predict.mvnormal}
\alias{predict.mvnormal}
\title{Predict Method for Multivariate Normal Distribution model}
\usage{
\method{predict}{mvnormal}(object, newdata = NULL)
}
\arguments{
\item{object}{An object of class `mvnormal`.}

\item{newdata}{New data for predictions in a form of matrix or conversible to matrix.  If NULL (default) predictions are given to training data.}
}
\value{
A vector of predictions.
}
\description{
This function predicts values using a `mvnormal` object.
}
