% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{print.kernel}
\alias{print.kernel}
\title{Print Method for Kernel}
\usage{
\method{print}{kernel}(x)
}
\arguments{
\item{x}{An object of class `kernel`.}
}
\description{
Print simple information for object of class `kernel`.
}
