% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R
\name{print.mahalanobis}
\alias{print.mahalanobis}
\title{Print Method for mahalanobis}
\usage{
\method{print}{mahalanobis}(model)
}
\arguments{
\item{model}{An object of class `mahalanobis`.}
}
\description{
Print simple information for object of class `mahalanobis`.
}
