% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promodel.R
\name{promodel}
\alias{promodel}
\title{Compute presence-only model}
\usage{
promodel(
  vals,
  method = "bioclim",
  na.rm = TRUE,
  dup.rm = FALSE,
  verbosity = 2,
  ...
)
}
\arguments{
\item{vals}{data frame with the values of the variables (columns) at the
presence localities (rows).}

\item{method}{method for computing the model. Options are `"bioclim"`
(default), `"convexhull"`, `"domain"`, `"mahalanobis"`, `"kernel"` or
`"mvnormal"`. Case is ignored and partial matching is allowed (e.g.
`method = "Mahal"`).}

\item{na.rm}{logical (default `TRUE`) indicating whether to remove rows with
missing values in `vals`. Otherwise, some methods may produce an error or
empty results.}

\item{dup.rm}{logical (default `FALSE`) indicating whether to remove rows
with duplicate values for all variables in `vals`.}

\item{verbosity}{numeric value indicating the amount of messages to display.
Default is 2 (maximum verbosity).}

\item{...}{Additional arguments passed to the underlying model functions,
such as \code{\link{bioclim}}, \code{\link{convexhull}},
\code{\link{kernel}}, or \code{\link{mvnormal}}.}
}
\value{
An object of class `proniche`, containing the fitted model in the
  element `proniche`. The internal structure of this element depends on the
  chosen `method`.
}
\description{
Compute true presence-only models based on environmental variables at
species occurrence sites, using a specified modeling method.
}
\details{
Implemented methods include:
\itemize{
  \item Bioclim (Nix, 1986; Booth et al., 2014)
  \item Convex hull in environmental space (using
    \code{\link[geometry:convhulln]{geometry::convhulln()}} and
    \code{\link[geometry:inhulln]{geometry::inhulln()}})
  \item Domain (Carpenter et al., 1993)
  \item Mahalanobis distance (Farber & Kadmon, 2003)
  \item Kernel density estimate (using \code{\link[ks:kde]{ks::kde()}})
  \item Multivariate normal distribution
}
}
\examples{
\dontrun{
library(terra)

# Predictor variables
elevation <- terra::rast(system.file("ex/elev.tif", package = "terra"))
slope <- terra::terrain(elevation, v = "slope")
roughness <- terra::terrain(elevation, v = "roughness")
vars <- c(elevation, slope, roughness)
terra::plot(vars)

# Random presence points
vars_coords <- terra::crds(vars)
set.seed(123)
pres_coords <- vars_coords[sample(1:nrow(vars_coords), 20), ]
vals <- terra::extract(vars, pres_coords)
head(vals)

# Compute and plot a model
mod <- promodel(vals, method = "mahalanobis")
plot(mod)
pred <- terra::predict(vars, mod)
terra::plot(pred)
}

}
\references{
Booth T.H., Nix H.A., Busby J.R. & Hutchinson M.F. (2014)
  bioclim: the first species distribution modelling package, its early
  applications and relevance to most current MaxEnt studies.
  *Diversity & Distributions*, 20(1):1–9.

Carpenter G., Gillison A.N. & Winter J. (1993)
  DOMAIN: a flexible modelling procedure for mapping potential distributions
  of plants and animals. *Biodiversity and Conservation*, 2(6):667–680.

Farber O. & Kadmon R. (2003)
  Assessment of alternative approaches for bioclimatic modeling with special
  emphasis on the Mahalanobis distance. *Ecological Modelling*,
  160(1–2):115–130.

Nix H.A. (1986)
  A biogeographic analysis of Australian elapid snakes.
  *Atlas of elapid snakes of Australia: Australian flora and fauna series 7*
  (ed. R. Longmore), pp. 4–15. Bureau of Flora and Fauna, Canberra.
}
\seealso{
\code{\link{quantReclass}};
  \code{\link[dismo:bioclim]{dismo::bioclim()}};
  \code{\link[predicts:envelope]{predicts::envelope()}};
  \code{\link[geometry:convhulln]{geometry::convhulln()}};
  \code{\link[ks:kde]{ks::kde()}}
}
\author{
Pedro Tarroso and A. Marcia Barbosa
}
\keyword{models}
\keyword{prediction}
\keyword{spatial}
